refactor to use supervisely.Applicataion instead of FastApi
