+++ .vscode/launch.json
@@ -10,7 +10,7 @@
             "request": "launch",
             "module": "uvicorn",
             "args": [
-                "src.main:g.app",
+                "src.main:app",
                 "--host",
                 "0.0.0.0",
                 "--port",
