--- src/choose_videos/widgets.py
@@ -1,1 +1,0 @@
-from supervisely.app.widgets import NotificationBox
