--- src/connect_to_model/widgets.py
@@ -1,7 +1,0 @@
-from supervisely.app.widgets import NotificationBox
-
-additional_tracking_required = NotificationBox(
-    title="Tracking Enabled",
-    description="All predictions will be processed with tracking algorithm",
-    widget_id="additional_tracking_required",
-)
