--- src/input_data/functions.py
@@ -1,1 +1,0 @@
-### some card functional
