--- src/input_data/widgets.py
@@ -1,3 +1,0 @@
-from supervisely.app.widgets import NotificationBox, SlyTqdm
-
-
