+++ src/ui/choose_videos/template.html
@@ -1,0 +1,65 @@
+<sly-card
+        title="Neural Network"
+        subtitle="Connect to deployed model"
+        :options="{collapsable: true}"
+
+        :disabled="{disabled: state.activeStep < 2, message: 'complete previous step'}"
+        ref="step2">
+    <div v-if="data.connected === false">
+        <div class="fflex">
+            <sly-select-app-session :group-id="data.teamId"
+                                    :app-session-id.sync="state.sessionId"
+                                    :options="data.ssOptions">
+            </sly-select-app-session>
+
+            <el-button
+                    style="margin-left: 10px"
+                    type="primary"
+                    class="regular"
+                    :loading="state.connectionLoading"
+                    :disabled="!state.sessionId"
+                    @click="state.connectionLoading=true; post('/connect-to-model/')"
+            >
+                <i class="zmdi zmdi-refresh mr10" v-if="!state.connectionLoading"></i>Connect to model
+            </el-button>
+
+        </div>
+        <sly-notification
+                v-if="data.connectionError.length > 0"
+                :options="{'name': 'Connection Error', 'type': 'error'}"
+                :content="data.connectionError">
+        </sly-notification>
+    </div>
+
+    <div v-if="data.connected === true" style="display: flex; flex-direction: row; justify-content: space-between">
+        <div>
+            <sly-field title="INFO: Successfully connected to model"
+                       description="Below are the parameters of the model"
+                       style="padding-top: 4px; padding-bottom: 6px">
+
+                <sly-icon slot="icon" :options="{ color: '#13ce66', bgColor: '#e1f7eb', rounded: false }">
+                    <i class="zmdi zmdi-circle"></i>
+                </sly-icon>
+            </sly-field>
+
+            <div style="margin: 22px 5px">
+                <h3>Model Stats:</h3>
+                <div v-for="v,k in data.modelInfo" class="mb10" style="word-break: break-all;">
+                    <div style="margin-bottom: 4px">
+                        <b>{{k}}</b>: <span>{{v}}</span>
+                    </div>
+                </div>
+            </div>
+
+
+            <el-button type="warning" class="regular mt10" :plain="true"
+                       @click="state.restartFrom = 2;"
+                       v-if="data.connected === true"
+                       :disabled="data.annotatingStarted">
+                <i class="zmdi zmdi-rotate-left mr5"></i> Connect to another model
+            </el-button>
+        </div>
+
+    </div>
+</sly-card>
+</sly-card>
