+++ src/ui/input_data/template.html
@@ -1,0 +1,43 @@
+<sly-card
+
+        title="Input data"
+        subtitle="Download videos from project"
+        :options="{collapsable: true}"
+        ref="step1"
+>
+    <!-- <sly-field title="" description="Project">
+        <a slot="title" target="_blank"
+           :href="`${data.instanceAddress}/projects/${data.projectId}/datasets`">{{data.projectName}} ({{data.projectItemsCount}}
+            videos)</a>
+        <sly-icon slot="icon" :options="{ imageUrl: `${data.projectPreviewUrl}` }"/>
+    </sly-field> -->
+    {{{widget._project_thumbnail}}}
+
+    <el-button
+            style="margin-top: 8px"
+            type="primary"
+            class="regular"
+            :loading="state.inputLoading"
+            @click="state.inputLoading=true; post('/select_projects_handler/')"
+            v-if="!data.done1 && data.projectItemsCount > 0">
+        Select
+    </el-button>
+
+
+
+    <sly-field v-if="data.projectItemsCount === 0" title="WARNING: No videos found in project"
+               description="please reselect input project"
+               style="margin-bottom: 0"
+    >
+        <sly-icon slot="icon" :options="{ color: '#fba607', bgColor: '#ffe9da', rounded: false }">
+            <i class="zmdi zmdi-alert-triangle"></i>
+        </sly-icon>
+    </sly-field>
+
+
+    <div v-if="data.done1" style="margin: 10px 0 0">
+        <i class="zmdi zmdi-check-circle mr5" style="color: #13ce66"></i>
+        <span style="color: #5a6772;">Project has been successfully selected</span>
+    </div>
+</sly-card>
+
